LayoutTests/media/track/track-webvtt-tc010-notimings.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            
 
            function trackLoaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "0");
                
                endTest();
            }

        </script>
    </head>
    <body>
    <p>Tests cue without timings are ignored.</p>
        <video>
            <track src="captions/tc010_notimings.vtt" onload="trackLoaded()">
        </video>
    </body>
</html>
