LayoutTests/media/track/track-webvtt-tc010-notimings-expected.txt

Tests cue without timings are ignored.

EXPECTED (video.textTracks[0].cues.length == '0') OK
END OF TEST
